Half-star ratings everywhere: 0.5 steps, tap halves, drag like a slider
LibraryItem
- rating goes from Int? to Double? (0.5 ... 5.0 in 0.5 steps).
  ratingValue and hasRating use Double now.

Firestore codec (LibraryRepository)
- encode writes the rating as a Double.
- decode accepts Double, legacy Int and NSNumber so existing
  documents — every rating you've already given — keep working
  without a migration.
- updateRating + markWatched signatures switch to Double?.

StarRatingView (interactive)
- Half-step granularity.
  Tap the left half of a star → that star becomes a half-star.
  Tap the right half → full star.
  Drag horizontally across the row → slider-style continuous rating.
- A second tap on the currently-selected half-step clears the rating
  back to 0 (so "Überspringen" still works in spirit).
- Mapping is per-star: left half = idx-0.5, right half = idx, with
  the gap rounding to the previous half-step so accidental taps in
  the seam don't jump a star.

StarRatingDisplay (read-only)
- Renders star.fill / star.leadinghalf.filled / star to show 0.5
  granularity for any Double rating.

Callsites refactored
- RatingOverlay drops its bespoke star block and uses StarRatingView
  with a Double @State (full slider + half-tap support inside the
  overlay). Commit fires when the binding flips to a non-zero value.
- LibraryDetailView, AddTitleSheet, DiscoverView, WatchlistView and
  DiscoverViewModel.toLibraryItem are all on Double? now.
- LibraryView grid cell picks up the half stars automatically via
  StarRatingDisplay.
- LibraryViewModel.minRating stays Int (we filter by whole stars) and
  compares against the new Double rating with `Double(minRating)`.

+++ Cineo/Features/Discover/DiscoverView.swift
@@ -330,7 +330,7 @@ struct DiscoverView: View {
         ratingCandidate = candidate
     }
 
-    private func commitRating(_ value: Int?, for candidate: DiscoverViewModel.Candidate) {
+    private func commitRating(_ value: Double?, for candidate: DiscoverViewModel.Candidate) {
         let item = viewModel.toLibraryItem(candidate, rating: value, watched: true)
         // Optimistic UI: dismiss the overlay and advance the stack right away.
         // The Firestore write runs in the background — first-time save latency
